Refactor cleaning of scrapping items
The logic for avoiding duplicates relays in the current year
that is being requested but as has been found the requested
year cannot be trusted anymore as the logic might change
constantly.

Instead of trying to clean scraped items a new solution
should strive to avoid duplicates at the moment of scraping.